refactor: merge Security page into Secrets, skip current password in dev mode
- Merge auth status + password management into the existing Secrets page
  (now titled "Security") instead of a separate /settings/security route
- Skip "Current Password" field when auth mode is disabled (dev mode),
  both frontend and backend
- Remove standalone /settings/security page and deduplicate sidebar entry
