+++ dashboard/src/app/settings/secrets/page.tsx
@@ -14,6 +14,8 @@ import {
   setSecret,
   deleteSecret,
   revealSecret,
+  getAuthStatus,
+  changePassword,
   type SecretsStatus,
   type EncryptionStatus,
   type SecretInfo,
@@ -35,12 +37,23 @@ import {
   CheckCircle,
   AlertCircle,
   Settings,
+  Clock,
+  AlertTriangle,
+  Info,
 } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { useToast } from '@/components/toast';
+import { toast } from '@/components/toast';
 
 export default function SecretsPage() {
   const { showError } = useToast();
+
+  // Fetch auth status
+  const { data: authStatus, isLoading: authLoading, mutate: mutateAuth } = useSWR(
+    'auth-status',
+    getAuthStatus,
+    { revalidateOnFocus: false }
+  );
 
   // Fetch encryption status (skill content encryption)
   const { data: encryptionStatus, isLoading: encryptionLoading } = useSWR(
@@ -84,6 +97,57 @@ export default function SecretsPage() {
 
   // Copy feedback
   const [copiedKey, setCopiedKey] = useState<string | null>(null);
+
+  // Password form state
+  const [currentPassword, setCurrentPassword] = useState('');
+  const [newPassword, setNewPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
+  const [savingPassword, setSavingPassword] = useState(false);
+
+  const isMultiUser = authStatus?.auth_mode === 'multi_user';
+  const isAuthDisabled = authStatus?.auth_mode === 'disabled';
+  const hasExistingPassword = authStatus?.password_source !== 'none';
+  const requireCurrentPassword = hasExistingPassword && !isAuthDisabled;
+
+  const handleChangePassword = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (newPassword !== confirmPassword) {
+      toast.error('Passwords do not match');
+      return;
+    }
+    if (newPassword.length < 8) {
+      toast.error('Password must be at least 8 characters');
+      return;
+    }
+    setSavingPassword(true);
+    try {
+      await changePassword({
+        current_password: requireCurrentPassword ? currentPassword : undefined,
+        new_password: newPassword,
+      });
+      toast.success('Password updated successfully');
+      setCurrentPassword('');
+      setNewPassword('');
+      setConfirmPassword('');
+      mutateAuth();
+    } catch (err) {
+      toast.error(err instanceof Error ? err.message : 'Failed to change password');
+    } finally {
+      setSavingPassword(false);
+    }
+  };
+
+  const passwordSourceLabel = {
+    dashboard: 'Dashboard-managed',
+    environment: 'Environment variable',
+    none: 'Not configured',
+  }[authStatus?.password_source ?? 'none'];
+
+  const authModeLabel = {
+    disabled: 'Disabled',
+    single_tenant: 'Single Tenant',
+    multi_user: 'Multi-User',
+  }[authStatus?.auth_mode ?? 'disabled'];
 
   // Private key management
   const [showKeyDialog, setShowKeyDialog] = useState(false);
@@ -298,7 +362,7 @@ export default function SecretsPage() {
     return type.replace(/_/g, ' ').replace(/\b\w/g, (c) => c.toUpperCase());
   };
 
-  const loading = encryptionLoading || secretsLoading;
+  const loading = authLoading || encryptionLoading || secretsLoading;
 
   if (loading) {
     return (
@@ -315,10 +379,165 @@ export default function SecretsPage() {
       <div className="max-w-4xl mx-auto space-y-8">
       {/* Header */}
       <div>
-        <h1 className="text-2xl font-semibold text-white mb-2">Encryption & Secrets</h1>
+        <h1 className="text-2xl font-semibold text-white mb-2">Security</h1>
         <p className="text-white/50">
-          Manage encryption for skill content and optional secret storage.
+          Authentication, encryption, and secrets management.
         </p>
+      </div>
+
+      {/* Auth Status Card */}
+      <div className="rounded-xl border border-white/[0.06] bg-white/[0.02] p-6">
+        <div className="flex items-center gap-3 mb-4">
+          <div className="flex h-9 w-9 items-center justify-center rounded-lg bg-white/[0.06]">
+            <Shield className="h-5 w-5 text-white/70" />
+          </div>
+          <div>
+            <h2 className="text-base font-medium text-white">Authentication</h2>
+            <p className="text-xs text-white/40">Current authentication configuration</p>
+          </div>
+        </div>
+
+        <div className="grid grid-cols-1 sm:grid-cols-2 gap-4 mb-6">
+          <div className="rounded-lg bg-white/[0.03] border border-white/[0.04] p-4">
+            <p className="text-xs text-white/40 mb-1">Auth Mode</p>
+            <span className={`inline-flex items-center gap-1.5 rounded-full px-2.5 py-0.5 text-xs font-medium ${
+              authStatus?.auth_mode === 'disabled'
+                ? 'bg-yellow-500/10 text-yellow-400'
+                : authStatus?.auth_mode === 'multi_user'
+                ? 'bg-blue-500/10 text-blue-400'
+                : 'bg-emerald-500/10 text-emerald-400'
+            }`}>
+              {authModeLabel}
+            </span>
+          </div>
+
+          <div className="rounded-lg bg-white/[0.03] border border-white/[0.04] p-4">
+            <p className="text-xs text-white/40 mb-1">Password Source</p>
+            <span className={`inline-flex items-center gap-1.5 rounded-full px-2.5 py-0.5 text-xs font-medium ${
+              authStatus?.password_source === 'dashboard'
+                ? 'bg-emerald-500/10 text-emerald-400'
+                : authStatus?.password_source === 'environment'
+                ? 'bg-blue-500/10 text-blue-400'
+                : 'bg-white/[0.06] text-white/50'
+            }`}>
+              {passwordSourceLabel}
+            </span>
+          </div>
+
+          {authStatus?.password_changed_at && (
+            <div className="rounded-lg bg-white/[0.03] border border-white/[0.04] p-4">
+              <p className="text-xs text-white/40 mb-1">Last Changed</p>
+              <p className="flex items-center gap-1.5 text-sm text-white/70">
+                <Clock className="h-3.5 w-3.5" />
+                {new Date(authStatus.password_changed_at).toLocaleString()}
+              </p>
+            </div>
+          )}
+
+          {authStatus?.dev_mode && (
+            <div className="rounded-lg bg-yellow-500/5 border border-yellow-500/10 p-4">
+              <p className="flex items-center gap-1.5 text-xs text-yellow-400">
+                <AlertTriangle className="h-3.5 w-3.5" />
+                Dev mode is enabled &mdash; authentication is bypassed
+              </p>
+            </div>
+          )}
+        </div>
+
+        {/* Password Management */}
+        {isMultiUser ? (
+          <div className="rounded-lg bg-white/[0.03] border border-white/[0.04] p-4">
+            <p className="flex items-center gap-2 text-sm text-white/50">
+              <Info className="h-4 w-4 shrink-0" />
+              In multi-user mode, passwords are managed via the <code className="mx-1 rounded bg-white/[0.06] px-1.5 py-0.5 text-xs font-mono">SANDBOXED_USERS</code> environment variable.
+            </p>
+          </div>
+        ) : (
+          <form onSubmit={handleChangePassword} className="space-y-4 max-w-md">
+            {!hasExistingPassword && (
+              <div className="rounded-lg bg-blue-500/5 border border-blue-500/10 p-3">
+                <p className="flex items-center gap-2 text-xs text-blue-400">
+                  <Info className="h-3.5 w-3.5 shrink-0" />
+                  No password is configured. Set one to enable authentication.
+                  {!authStatus?.dev_mode && ' You will also need JWT_SECRET set as an environment variable.'}
+                </p>
+              </div>
+            )}
+
+            {requireCurrentPassword && (
+              <div>
+                <label className="block text-sm font-medium text-white/70 mb-1.5">
+                  Current Password
+                </label>
+                <input
+                  type="password"
+                  value={currentPassword}
+                  onChange={(e) => setCurrentPassword(e.target.value)}
+                  className="w-full rounded-lg border border-white/[0.08] bg-white/[0.04] px-3 py-2 text-sm text-white placeholder-white/30 focus:border-white/20 focus:outline-none focus:ring-1 focus:ring-white/20"
+                  placeholder="Enter current password"
+                  required
+                />
+              </div>
+            )}
+
+            <div>
+              <label className="block text-sm font-medium text-white/70 mb-1.5">
+                New Password
+              </label>
+              <input
+                type="password"
+                value={newPassword}
+                onChange={(e) => setNewPassword(e.target.value)}
+                className="w-full rounded-lg border border-white/[0.08] bg-white/[0.04] px-3 py-2 text-sm text-white placeholder-white/30 focus:border-white/20 focus:outline-none focus:ring-1 focus:ring-white/20"
+                placeholder="At least 8 characters"
+                minLength={8}
+                required
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-white/70 mb-1.5">
+                Confirm New Password
+              </label>
+              <input
+                type="password"
+                value={confirmPassword}
+                onChange={(e) => setConfirmPassword(e.target.value)}
+                className="w-full rounded-lg border border-white/[0.08] bg-white/[0.04] px-3 py-2 text-sm text-white placeholder-white/30 focus:border-white/20 focus:outline-none focus:ring-1 focus:ring-white/20"
+                placeholder="Confirm new password"
+                minLength={8}
+                required
+              />
+              {confirmPassword && newPassword !== confirmPassword && (
+                <p className="mt-1 text-xs text-red-400">Passwords do not match</p>
+              )}
+            </div>
+
+            <button
+              type="submit"
+              disabled={savingPassword || !newPassword || newPassword !== confirmPassword || newPassword.length < 8}
+              className="inline-flex items-center gap-2 rounded-lg bg-white/[0.08] px-4 py-2 text-sm font-medium text-white transition-colors hover:bg-white/[0.12] disabled:opacity-40 disabled:cursor-not-allowed"
+            >
+              {savingPassword ? (
+                <>
+                  <span className="h-4 w-4 animate-spin rounded-full border-2 border-white/20 border-t-white/70" />
+                  Saving...
+                </>
+              ) : (
+                <>
+                  <Lock className="h-4 w-4" />
+                  {hasExistingPassword ? 'Update Password' : 'Set Password'}
+                </>
+              )}
+            </button>
+
+            {authStatus?.password_source === 'environment' && (
+              <p className="text-xs text-white/30">
+                Setting a dashboard password will take priority over the DASHBOARD_PASSWORD environment variable.
+              </p>
+            )}
+          </form>
+        )}
       </div>
 
       {/* Encryption Status Card */}
